+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/OtlpExporterOptions.cs
@@ -98,10 +98,9 @@ namespace OpenTelemetry.Exporter
         }
 
         /// <summary>
-        /// Gets or sets the target to which the exporter is going to send traces.
-        /// Must be a valid Uri with scheme (http) and host, and
-        /// may contain a port and path. Secure connection(https) is not
-        /// supported.
+        /// Gets or sets the target to which the exporter is going to send telemetry.
+        /// Must be a valid Uri with scheme (http or https) and host, and
+        /// may contain a port and path. The default value is http://localhost:4317.
         /// </summary>
         public Uri Endpoint { get; set; } = new Uri("http://localhost:4317");
 
@@ -112,7 +111,7 @@ namespace OpenTelemetry.Exporter
         public string Headers { get; set; }
 
         /// <summary>
-        /// Gets or sets the max waiting time (in milliseconds) for the backend to process each span batch. The default value is 10000.
+        /// Gets or sets the max waiting time (in milliseconds) for the backend to process each batch. The default value is 10000.
         /// </summary>
         public int TimeoutMilliseconds { get; set; } = 10000;